requirements.txt
pyyaml
aiogram
loguru
asyncpg